Fix move to right package
